feature(readd dependencies) re-add dependencies that were removed [Starts #165922868]
