+++ package.json
@@ -4,13 +4,13 @@
   "description": "Quick Credit is an online lending platform that provides short term soft loans to individuals.",
   "main": "server.js",
   "scripts": {
-    "start": "cross-env NODE_ENV=development nodemon --exec babel-node src/server.js",
-    "dev": "cross-env NODE_ENV=development nodemon --exec babel-node src/server.js",
+    "start": " nodemon --exec babel-node src/server.js",
+    "dev": " nodemon --exec babel-node src/server.js",
     "build": "babel src --out-dir dist",
     "serve": "node dist/server.js",
-    "test": "cross-env NODE_ENV=development nyc mocha -r @babel/register ./src/tests/test",
-    "test:repl": "cross-env NODE_ENV=development mocha debug -r @babel/register ./src/tests/test",
-    "test:debug": "cross-env NODE_ENV=development mocha --inspect --debug-brk",
+    "test": " nyc mocha -r  @babel/register --recursive ./src/test/*.test.js",
+    "test:repl": " mocha debug -r @babel/register  ./src/tests",
+    "test:debug": " mocha --inspect --debug-brk",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
     "coveralls": "nyc --reporter=lcov --reporter=text-lcov npm test"
   },
@@ -55,6 +55,7 @@
     "dotenv": "^8.0.0",
     "express": "^4.16.4",
     "express-sessions": "^1.0.6",
+    "express-validator": "^5.3.1",
     "jsonwebtoken": "^8.5.1",
     "node": "^12.1.0",
     "supertest": "^4.0.2",
