+++ src/server.js
@@ -2,25 +2,26 @@
 import dotenv from 'dotenv';
 import express from 'express';
 import path from 'path';
+import validator from 'express-validator';
 import volleyball from 'volleyball';
 import bodyParser from 'body-parser';
-import config from './config/config';
+// import config from './config/config';
 import routes from './routes/main';
 import apiRoutes from './routes/api/routes';
 
 dotenv.config();
 const app = express();
 
-const environment = process.env.NODE_ENV; // development
-const stage = config[environment];
+const port = process.env.PORT || 4000;
 
-if (environment !== 'production') {
+if (port !== 'production') {
   app.use(volleyball);
 }
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
   extended: true,
 }));
+app.use(validator());
 // the API routes
 app.use(express.static(path.join('UI')));
 app.use('/', routes);
@@ -37,6 +38,6 @@ app.use((err, req, res) => {
   res.status(err.status || 500);
 });
 
-app.listen(`${stage.port}`, () => console.log(` app listening on port ${stage.port}!`));
+app.listen(`${port}`, () => console.log(` app listening on port ${port}!`));
 
 module.exports = app;
